Remove Kinto from vendors and move sendMetrics to Utils.js to be shareable
